ecotrack/python-backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
opencv-python>=4.5.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.0.0
